refactor: StoreServiceTest와 PointServiceTest 분리, StoreServiceTest는 통합테스트로 동시성 테스트를 진행하며 PointServiceTest는 transfer 테스트만 진행하는 단위 테스트 (#24)
