+++ src/main/java/com/advent/backend/event/PurchaseEvent.java
@@ -1,0 +1,5 @@
+package com.advent.backend.event;
+
+import com.advent.backend.entity.Member;
+
+public record PurchaseEvent(Member sender, Member receiver, String itemName) {}
