--- src/main/java/com/advent/backend/event/TransferEvent.java
@@ -1,6 +1,0 @@
-package com.advent.backend.event;
-
-import com.advent.backend.entity.Member;
-import java.util.UUID;
-
-public record TransferEvent(UUID txId, Member sender, Member receiver, String itemName) {}
